--- integrations/azure/azure_integration/exceptions.py
@@ -1,9 +1,0 @@
-from port_ocean.exceptions.base import BaseOceanException
-
-
-class AzureIntegrationException(BaseOceanException):
-    pass
-
-
-class AzureIntegrationNotFoundKindInPortAppConfig(AzureIntegrationException):
-    pass
