+++ integrations/azure/pyproject.toml
@@ -12,7 +12,6 @@ azure-mgmt-resource = "23.0.1"
 azure-identity = "^1.13.0"
 aiohttp = "^3.7.4"
 cloudevents = "^1.9.0"
-requests = "^2.31.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.2"
